+++ resolvers/webpack/CHANGELOG.md
@@ -6,6 +6,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 ## Unreleased
 ### Added
 - Support for explicit Webpack config object in `.eslintrc.*`. ([#572], thanks [@jameslnewell])
+- Added `resolve.modules` to configs for webpack2 support ([#569], thanks [@toshafed])
 
 ## 0.6.0 - 2016-09-13
 ### Added
@@ -65,6 +66,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
   Thanks to [@gausie] for the initial PR ([#164], ages ago! 😅) and [@jquense] for tests ([#278]).
 
 [#572]: https://github.com/benmosher/eslint-plugin-import/pull/572
+[#569]: https://github.com/benmosher/eslint-plugin-import/pull/569
 [#533]: https://github.com/benmosher/eslint-plugin-import/pull/533
 [#413]: https://github.com/benmosher/eslint-plugin-import/pull/413
 [#377]: https://github.com/benmosher/eslint-plugin-import/pull/377
@@ -92,3 +94,4 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 [@Kovensky]: https://github.com/Kovensky
 [@grahamb]: https://github.com/grahamb
 [@jameslnewell]: https://github.com/jameslnewell
+[@toshafed]: https://github.com/toshafed